renamed ServiceCard to ServicesCard
