+++ src/routes/[lang=lang]/services/+page.svelte
@@ -1,7 +1,7 @@
 <script>
   import { LL, locale } from '$i18n/i18n-svelte';
   import Hero from '$lib/Hero.svelte';
-  import ServiceCard from './ServiceCard.svelte';
+  import ServicesCard from './ServicesCard.svelte';
 
   const services =
     /** @type {Array.<keyof import('$i18n/i18n-types').Translation['services']['all']>} */
@@ -20,7 +20,7 @@
 <main class="flex flex-row flex-wrap px-2 py-8 justify-center max-w-6xl mx-auto">
   {#each services as service}
     <div class="sm:w-1/2 lg:w-1/3 p-1">
-      <ServiceCard
+      <ServicesCard
         title={$LL.services.all[service].title()}
         teaser={$LL.services.all[service].teaser()}
         link="/{$locale}/services/{service}"
